+++ changelog.md
@@ -1,1 +1,1 @@
-- Separate converter and incubator screens
+- Fixed farmer and iron farm not being able to output items